Start move to unified query executor
