+++ tsdb/query_executor.go
@@ -27,7 +27,6 @@ type QueryExecutor struct {
 		Authenticate(username, password string) (*meta.UserInfo, error)
 		RetentionPolicy(database, name string) (rpi *meta.RetentionPolicyInfo, err error)
 		UserCount() (int, error)
-
 		ShardGroupsByTimeRange(database, policy string, min, max time.Time) (a []meta.ShardGroupInfo, err error)
 		NodeID() uint64
 	}
@@ -35,6 +34,11 @@ type QueryExecutor struct {
 	// Executes statements relating to meta data.
 	MetaStatementExecutor interface {
 		ExecuteStatement(stmt influxql.Statement) *influxql.Result
+	}
+
+	// Maps shards for queries.
+	ShardMapper interface {
+		CreateMapper(shardID uint64, stmt string, chunkSize int) (Mapper, error)
 	}
 
 	Logger *log.Logger
@@ -200,6 +204,59 @@ func (q *QueryExecutor) ExecuteQuery(query *influxql.Query, database string, chu
 	return results, nil
 }
 
+// Plan creates an execution plan for the given SelectStatement and returns an Executor.
+func (q *QueryExecutor) plan(stmt *influxql.SelectStatement, chunkSize int) (Executor, error) {
+	shards := map[uint64]meta.ShardInfo{} // Shards requiring mappers.
+
+	for _, src := range stmt.Sources {
+		mm, ok := src.(*influxql.Measurement)
+		if !ok {
+			return nil, fmt.Errorf("invalid source type: %#v", src)
+		}
+
+		// Replace instances of "now()" with the current time, and check the resultant times.
+		stmt.Condition = influxql.Reduce(stmt.Condition, &influxql.NowValuer{Now: time.Now().UTC()})
+		tmin, tmax := influxql.TimeRange(stmt.Condition)
+		if tmax.IsZero() {
+			tmax = time.Now()
+		}
+		if tmin.IsZero() {
+			tmin = time.Unix(0, 0)
+		}
+
+		// Build the set of target shards. Using shard IDs as keys ensures each shard ID
+		// occurs only once.
+		shardGroups, err := q.MetaStore.ShardGroupsByTimeRange(mm.Database, mm.RetentionPolicy, tmin, tmax)
+		if err != nil {
+			return nil, err
+		}
+		for _, g := range shardGroups {
+			for _, sh := range g.Shards {
+				shards[sh.ID] = sh
+			}
+		}
+	}
+
+	// Build the Mappers, one per shard.
+	mappers := []Mapper{}
+	for _, sh := range shards {
+		m, err := q.ShardMapper.CreateMapper(sh.ID, stmt.String(), chunkSize)
+		if err != nil {
+			return nil, err
+		}
+		if m == nil {
+			// No data for this shard, skip it.
+			continue
+		}
+		mappers = append(mappers, m)
+	}
+
+	if stmt.IsRawQuery && !stmt.HasDistinct() {
+		return NewRawExecutor(stmt, mappers, chunkSize), nil
+	}
+	return NewAggregateExecutor(stmt, mappers), nil
+}
+
 // executeSelectStatement plans and executes a select statement against a database.
 func (q *QueryExecutor) executeSelectStatement(statementID int, stmt *influxql.SelectStatement, results chan *influxql.Result, chunkSize int) error {
 	// Perform any necessary query re-writing.
@@ -209,9 +266,7 @@ func (q *QueryExecutor) executeSelectStatement(statementID int, stmt *influxql.S
 	}
 
 	// Plan statement execution.
-	p := NewPlanner(q.store)
-	p.MetaStore = q.MetaStore
-	e, err := p.Plan(stmt, chunkSize)
+	e, err := q.plan(stmt, chunkSize)
 	if err != nil {
 		return err
 	}
